// src/app/models/promocode-batch-details.ts
export class PromoCodeBatchDetails {
    batchId: string;
    batchName: string;
    brandId: number;
    brandName: string;
    loyaltyPoints: string;
    expirationDateTime: string;
    noOfPromoCodes: number;
    noOfPromoCodesScannedSuccessfully: number;
    isExpired: boolean;
    isDeleted: string;
    createdOn: string;
    createdByUserId: string;
    filteredCount: string;
}